Module Obatcher.Make_Poly

Parameters

module S : Service_Poly

Signature

type 'a t

t represents the service now abstracted with batching made implicit & automatic

val init : ?cfg:S.cfg -> unit -> 'a t

init ?cfg () initializes service with implicit batching. cfg is passed on the service initialization

val exec : 'a t -> ('a, 'b) S.op -> 'b

exec t op is the API call for a singular operation on the service with operations being automatically batched before passed to the service

val get_internal : 'a t -> 'a S.t

get_internal t extracts underlying internal service

alert unsafe Should not be used typically by application